pyproject.toml
[project]
name = "softwareproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "gunicorn>=25.1.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]
